+++ scripts/Publish.ps1
@@ -1,6 +1,6 @@
 param(
     [ValidatePattern('^\d+\.\d+\.\d+$')]
-    [string]$Version = '0.7.0',
+    [string]$Version = '0.7.1',
 
     [string]$CertificateThumbprint = '',